Introduce common config to share configs
