+++ modules/launcher/src/main/resources/broker.yaml
@@ -14,7 +14,23 @@
 # specific language governing permissions and limitations
 # under the License.
 
-# Broker Configurations.
+# Broker configurations
+ballerina.broker:
+ # When the in-memory mode is enabled all message and context data including durable entities are held in memory.
+ # Please note that this can increase memory usage of the broker. Therefore it is advised to allocate more memory in
+ # the JVM which run the broker.
+ enableInMemoryMode: false
+
+ # Datasource configurations used to communicate with the database.
+ dataSource:
+  # Database URL.
+  url: jdbc:h2:./database/MB_DB
+  # Database username.
+  user: wso2carbon
+  # Database password.
+  password: wso2carbon
+
+# Broker core configurations.
 ballerina.broker.core:
  # Maximum number of messages kept in a non-durable queue. Increasing this number can increase the memory consumption.
  # Messages published to queue are dropped after this limit is reached.
@@ -31,15 +47,6 @@ ballerina.broker.core:
   # The time that the delivery task will wait when the queue is empty or no consumers are available for message
   # delivery in milliseconds.
   idleTaskDelay: 50
-
- # Datasource configurations used to communicate with the database.
- dataSource:
-  # Database URL.
-  url: jdbc:h2:./database/MB_DB
-  # Database username.
-  user: wso2carbon
-  # Database password.
-  password: wso2carbon
 
 # Broker auth related configurations.
 ballerina.broker.auth:
